New icon
New icon for notes so that we can merge the database into Mishal Lab
