--- lib/all_screen.dart
@@ -1,8 +1,0 @@
-import 'package:mishallab_app/accounts_page.dart';
-
-import 'package:flutter/material.dart';
-
-void main() {
-  runApp(MyApp());
-}
-
