+++ lib/chat_page.dart
@@ -1,20 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:mishallab_app/accounts_page.dart';
 
-void main() {
-  runApp(MyApp());
-}
 
-class ChatScreen extends StatelessWidget {
-  const ChatScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Text('Chat Screen'),
-    );
-  }
-}
 // ignore: use_key_in_widget_constructors
 class ChatPage extends StatelessWidget {
   @override
